fix: add command  an event list
