+++ src/deploy-events.js
@@ -1,16 +1,4 @@
-const guildScheduledEventCreate = require('./events/guildScheduledEventCreate');
-const interactionCreateEvent = require('./events/interactionCreate');
-const pollVotationResultEvent = require('./events/pollVotationResults');
-const qaMetionEvent = require('./events/qaMention');
-const readyEvent = require('./events/ready');
-
-const eventList = [
-  guildScheduledEventCreate,
-  interactionCreateEvent,
-  pollVotationResultEvent,
-  qaMetionEvent,
-  readyEvent,
-];
+const eventList = require('./utils/eventList');
 
 module.exports = (client) => {
   for (const event of eventList) {
